Frame pour multithread

+++ main.py
@@ -1,6 +1,11 @@
 import csv, os, time, copy
 from multi import *
 from single import *
+from mpi4py import MPI as mpi
+
+comm = mpi.COMM_WORLD
+size = comm.size
+rank = comm.rank
 
 sudokus = []
 filenameMulti = "outputMulti.txt"
@@ -15,44 +20,54 @@ def write_sudoku(board,file):
             elif i % 3 == 2:
                 f.write("|" + "---"*7 + "--|\n")
 
-
-if os.path.exists(filenameMulti): os.remove(filenameMulti)
-if os.path.exists(filenameSingle): os.remove(filenameSingle)
-with open('input.txt', 'r') as f:
-    current_sudoku = []
-    row = []
-    for line in f:
-        for num in line.split(','):
-            if num != '\n':
-                row.append(int(num))
-        if row != []:
-            current_sudoku.append(row)
-        if len(current_sudoku) == 9:
-            write_sudoku(current_sudoku,filenameMulti)
-            write_sudoku(current_sudoku,filenameSingle)
-            sudokus.append(current_sudoku)
-            current_sudoku = []
+if rank == 0:
+    if os.path.exists(filenameMulti): os.remove(filenameMulti)
+    if os.path.exists(filenameSingle): os.remove(filenameSingle)
+    with open('input.txt', 'r') as f:
+        current_sudoku = []
         row = []
+        for line in f:
+            for num in line.split(','):
+                if num != '\n':
+                    row.append(int(num))
+            if row != []:
+                current_sudoku.append(row)
+            if len(current_sudoku) == 9:
+                write_sudoku(current_sudoku,filenameMulti)
+                write_sudoku(current_sudoku,filenameSingle)
+                sudokus.append(current_sudoku)
+                current_sudoku = []
+            row = []
 
 #single
 sudokusCopy = copy.deepcopy(sudokus)
-tStart = time.time()
-finishedSudokus = sudokuSolverS(sudokus)
-tEnd = time.time()
-finalTime = tEnd - tStart
-with open(filenameSingle, 'a') as f:
-        f.write("Temps : "+ str(finalTime) +"\n")
 
-for current_sudoku in finishedSudokus :
-    write_sudoku(current_sudoku,filenameSingle)
+if rank == 0:
+    tStart = time.time()
+    finishedSudokus = sudokuSolverS(sudokus)
+    tEnd = time.time()
+    finalTime = tEnd - tStart
+    with open(filenameSingle, 'a') as f:
+            f.write("Temps : "+ str(finalTime) +"\n")
+
+    for current_sudoku in finishedSudokus :
+        write_sudoku(current_sudoku,filenameSingle)
 
 #multi
-tStart = time.time()
-finishedSudokus = sudokuSolverM(sudokusCopy)
-tEnd = time.time()
-finalTime = tEnd - tStart
-with open(filenameMulti, 'a') as f:
-        f.write("Temps : "+ str(finalTime) +"\n")
+if rank == 0:
+    tStart = time.time()
 
-for current_sudoku in finishedSudokus :
-    write_sudoku(current_sudoku,filenameMulti)
+comm.barrier()
+
+finishedSudokus = sudokuSolverM(sudokusCopy)
+
+comm.barrier()
+
+if rank == 0:
+    tEnd = time.time()
+    finalTime = tEnd - tStart
+    with open(filenameMulti, 'a') as f:
+            f.write("Temps : "+ str(finalTime) +"\n")
+
+    for current_sudoku in finishedSudokus :
+        write_sudoku(current_sudoku,filenameMulti)
